fastapi/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
sounddevice==0.4.6
scipy==1.12.0
numpy==1.26.4
deepgram-sdk==3.1.6
google-generativeai==0.3.2
yt-dlp==2024.3.10
moviepy==1.0.3